+++ example/example.dart
@@ -103,7 +103,7 @@ Future<void> main() async {
     );
 
     // High-performance Volume Stream endpoint is available.
-    final volumeStream = await twitter.tweetsService.connectVolumeStream();
+    final volumeStream = await twitter.tweetsService.connectSampleStream();
     await for (final response in volumeStream.stream.handleError(print)) {
       print(response);
     }