+++ lib/src/service/tweets/tweet_data.freezed.dart
@@ -223,6 +223,13 @@ mixin _$TweetData {
   /// other app.
   String? get source => throw _privateConstructorUsedError;
 
+  /// Indicates if a Tweet is eligible for edit, how long it is editable for,
+  /// and the number of remaining edits.
+  ///
+  /// Editable Tweets can be edited for the first 30 minutes after
+  /// creation and can be edited up to five times.
+  TweetEditControls? get editControls => throw _privateConstructorUsedError;
+
   /// Contains withholding details for [withheld content](https://help.twitter.com/en/rules-and-policies/tweet-withheld-by-country).
   ///
   /// To return this field, add `TweetField.withheld` in the request's query
@@ -269,6 +276,7 @@ abstract class $TweetDataCopyWith<$Res> {
       @JsonKey(name: 'possibly_sensitive') bool? isPossiblySensitive,
       @JsonKey(name: 'reply_settings') ReplySetting? replySetting,
       String? source,
+      TweetEditControls? editControls,
       TweetWithheld? withheld,
       DateTime? createdAt});
 
@@ -279,6 +287,7 @@ abstract class $TweetDataCopyWith<$Res> {
   $PromotedTweetMetricsCopyWith<$Res>? get promotedMetrics;
   $PublicTweetMetricsCopyWith<$Res>? get publicMetrics;
   $GeoCopyWith<$Res>? get geo;
+  $TweetEditControlsCopyWith<$Res>? get editControls;
   $TweetWithheldCopyWith<$Res>? get withheld;
 }
 
@@ -310,6 +319,7 @@ class _$TweetDataCopyWithImpl<$Res> implements $TweetDataCopyWith<$Res> {
     Object? isPossiblySensitive = freezed,
     Object? replySetting = freezed,
     Object? source = freezed,
+    Object? editControls = freezed,
     Object? withheld = freezed,
     Object? createdAt = freezed,
   }) {
@@ -386,6 +396,10 @@ class _$TweetDataCopyWithImpl<$Res> implements $TweetDataCopyWith<$Res> {
           ? _value.source
           : source // ignore: cast_nullable_to_non_nullable
               as String?,
+      editControls: editControls == freezed
+          ? _value.editControls
+          : editControls // ignore: cast_nullable_to_non_nullable
+              as TweetEditControls?,
       withheld: withheld == freezed
           ? _value.withheld
           : withheld // ignore: cast_nullable_to_non_nullable
@@ -476,6 +490,17 @@ class _$TweetDataCopyWithImpl<$Res> implements $TweetDataCopyWith<$Res> {
   }
 
   @override
+  $TweetEditControlsCopyWith<$Res>? get editControls {
+    if (_value.editControls == null) {
+      return null;
+    }
+
+    return $TweetEditControlsCopyWith<$Res>(_value.editControls!, (value) {
+      return _then(_value.copyWith(editControls: value));
+    });
+  }
+
+  @override
   $TweetWithheldCopyWith<$Res>? get withheld {
     if (_value.withheld == null) {
       return null;
@@ -512,6 +537,7 @@ abstract class _$$_TweetDataCopyWith<$Res> implements $TweetDataCopyWith<$Res> {
       @JsonKey(name: 'possibly_sensitive') bool? isPossiblySensitive,
       @JsonKey(name: 'reply_settings') ReplySetting? replySetting,
       String? source,
+      TweetEditControls? editControls,
       TweetWithheld? withheld,
       DateTime? createdAt});
 
@@ -529,6 +555,8 @@ abstract class _$$_TweetDataCopyWith<$Res> implements $TweetDataCopyWith<$Res> {
   $PublicTweetMetricsCopyWith<$Res>? get publicMetrics;
   @override
   $GeoCopyWith<$Res>? get geo;
+  @override
+  $TweetEditControlsCopyWith<$Res>? get editControls;
   @override
   $TweetWithheldCopyWith<$Res>? get withheld;
 }
@@ -563,6 +591,7 @@ class __$$_TweetDataCopyWithImpl<$Res> extends _$TweetDataCopyWithImpl<$Res>
     Object? isPossiblySensitive = freezed,
     Object? replySetting = freezed,
     Object? source = freezed,
+    Object? editControls = freezed,
     Object? withheld = freezed,
     Object? createdAt = freezed,
   }) {
@@ -639,6 +668,10 @@ class __$$_TweetDataCopyWithImpl<$Res> extends _$TweetDataCopyWithImpl<$Res>
           ? _value.source
           : source // ignore: cast_nullable_to_non_nullable
               as String?,
+      editControls: editControls == freezed
+          ? _value.editControls
+          : editControls // ignore: cast_nullable_to_non_nullable
+              as TweetEditControls?,
       withheld: withheld == freezed
           ? _value.withheld
           : withheld // ignore: cast_nullable_to_non_nullable
@@ -673,6 +706,7 @@ class _$_TweetData implements _TweetData {
       @JsonKey(name: 'possibly_sensitive') this.isPossiblySensitive,
       @JsonKey(name: 'reply_settings') this.replySetting,
       this.source,
+      this.editControls,
       this.withheld,
       this.createdAt})
       : _referencedTweets = referencedTweets,
@@ -932,6 +966,14 @@ class _$_TweetData implements _TweetData {
   @override
   final String? source;
 
+  /// Indicates if a Tweet is eligible for edit, how long it is editable for,
+  /// and the number of remaining edits.
+  ///
+  /// Editable Tweets can be edited for the first 30 minutes after
+  /// creation and can be edited up to five times.
+  @override
+  final TweetEditControls? editControls;
+
   /// Contains withholding details for [withheld content](https://help.twitter.com/en/rules-and-policies/tweet-withheld-by-country).
   ///
   /// To return this field, add `TweetField.withheld` in the request's query
@@ -953,7 +995,7 @@ class _$_TweetData implements _TweetData {
 
   @override
   String toString() {
-    return 'TweetData(id: $id, text: $text, authorId: $authorId, inReplyToUserId: $inReplyToUserId, conversationId: $conversationId, referencedTweets: $referencedTweets, contextAnnotations: $contextAnnotations, entities: $entities, attachments: $attachments, privateMetrics: $privateMetrics, organicMetrics: $organicMetrics, promotedMetrics: $promotedMetrics, publicMetrics: $publicMetrics, geo: $geo, lang: $lang, isPossiblySensitive: $isPossiblySensitive, replySetting: $replySetting, source: $source, withheld: $withheld, createdAt: $createdAt)';
+    return 'TweetData(id: $id, text: $text, authorId: $authorId, inReplyToUserId: $inReplyToUserId, conversationId: $conversationId, referencedTweets: $referencedTweets, contextAnnotations: $contextAnnotations, entities: $entities, attachments: $attachments, privateMetrics: $privateMetrics, organicMetrics: $organicMetrics, promotedMetrics: $promotedMetrics, publicMetrics: $publicMetrics, geo: $geo, lang: $lang, isPossiblySensitive: $isPossiblySensitive, replySetting: $replySetting, source: $source, editControls: $editControls, withheld: $withheld, createdAt: $createdAt)';
   }
 
   @override
@@ -990,6 +1032,8 @@ class _$_TweetData implements _TweetData {
             const DeepCollectionEquality()
                 .equals(other.replySetting, replySetting) &&
             const DeepCollectionEquality().equals(other.source, source) &&
+            const DeepCollectionEquality()
+                .equals(other.editControls, editControls) &&
             const DeepCollectionEquality().equals(other.withheld, withheld) &&
             const DeepCollectionEquality().equals(other.createdAt, createdAt));
   }
@@ -1016,6 +1060,7 @@ class _$_TweetData implements _TweetData {
         const DeepCollectionEquality().hash(isPossiblySensitive),
         const DeepCollectionEquality().hash(replySetting),
         const DeepCollectionEquality().hash(source),
+        const DeepCollectionEquality().hash(editControls),
         const DeepCollectionEquality().hash(withheld),
         const DeepCollectionEquality().hash(createdAt)
       ]);
@@ -1056,6 +1101,7 @@ abstract class _TweetData implements TweetData {
       @JsonKey(name: 'reply_settings')
           final ReplySetting? replySetting,
       final String? source,
+      final TweetEditControls? editControls,
       final TweetWithheld? withheld,
       final DateTime? createdAt}) = _$_TweetData;
 
@@ -1280,6 +1326,14 @@ abstract class _TweetData implements TweetData {
   /// - Determine if a Twitter user posted from the web, mobile device, or
   /// other app.
   String? get source;
+  @override
+
+  /// Indicates if a Tweet is eligible for edit, how long it is editable for,
+  /// and the number of remaining edits.
+  ///
+  /// Editable Tweets can be edited for the first 30 minutes after
+  /// creation and can be edited up to five times.
+  TweetEditControls? get editControls;
   @override
 
   /// Contains withholding details for [withheld content](https://help.twitter.com/en/rules-and-policies/tweet-withheld-by-country).