+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v4.1.5
+
+- Added `TweetEditControls` object. This object holds information about the editing of tweets. You can use it from `TweetData.editControls`, and also you must pass the `editControls` element to `tweetFields`. ([#472](https://github.com/twitter-dart/twitter-api-v2/issues/472))
+
 ## v4.1.4
 
 - Renamed object from `ComplianceData` to `BatchComplianceData`. If you are using the object name `ComplianceData` explicitly, modify it to `BatchComplianceData`. ([#462](https://github.com/twitter-dart/twitter-api-v2/issues/462))